Fixed:   Scoreboard crash on nil   Scoreboard misaligned columns   Semifinal submission allowed by non-semis team   Index formatting in admin   Bad formatting of open teams in admin   Bad formatting of open teams in verify team data   Poor data for location of open teams   Diff errors on some pair case and missing html_safe   Home page spacing   Do-submit eval() usage -> direct DOM add/delete for do_submit

+++ db/seeds.rb
@@ -2,7 +2,7 @@
 Administrator.delete_all
 Administrator.find_or_create_by_name(:name => 'admin', :password => 'foobarbaz', :password_confirmation => 'foobarbaz')
 # For development, use factory girl to build some records.
-if Rails.env.development?
+if false && Rails.env.development?
   REDIS.flushall
   Team.delete_all
   Member.delete_all
